app/requirements.txt
beautifulsoup4==4.11.1
boto3==1.18.44
botocore==1.21.44
icecream==2.1.1
pandas==1.3.3
psycopg2_binary==2.9.3
selenium==4.4.0
